requirements.txt
alembic==0.9.1
elasticsearch==5.4.0
eventlet==0.21.0
flask-alchemydumps==0.0.10
flask-bcrypt==0.7.1
flask-login==0.4.0
flask-migrate==2.1.1
flask-script==2.0.6
flask-session==0.3.1
flask-sqlalchemy==2.3.2
flask==0.12.2
google-api-python-client==1.6.4
gunicorn
jinja2==2.9.6
psycopg2==2.7.3.2
pyjwt==1.5.3
python-dateutil==2.6.1
pytz==2017.3
simplejson==3.12.0
sqlalchemy==1.1.15
uwsgi==2.0.15